requirements.txt
Flask==3.1.3
Flask-SQLAlchemy==3.1.1
openai>=2.26.0
python-dotenv==1.2.2
pytest==9.0.2
pypdf==6.5.0
python-docx==1.2.0
requests==2.32.5
